minor fixes restruct
