+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: nchennaf <nchennaf@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/22 13:25:30 by nchennaf          #+#    #+#             */
-/*   Updated: 2022/08/11 16:48:08 by nchennaf         ###   ########.fr       */
+/*   Updated: 2022/08/11 17:01:52 by nchennaf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,17 +50,10 @@ typedef struct s_builtins
 } t_builtins;
 
 //builtins.c
-void	b_init(t_builtins *bs, int argc, char *argv[], char *envp[]);
 int		b_pwd(t_builtins *bs);
 int		b_cd(t_builtins *bs);
 int		b_exit(t_builtins *bs);
 int		b_env(t_builtins *bs);
-
-
-
-
-void	malloc_checker(char *s);
-
 
 //env_utils.c
 char	**parse_env(char *s);
@@ -69,12 +62,10 @@ void	print_env(char **elem);
 
 //b_export.c
 int		b_export(t_builtins *bs);
-void	add_key(t_builtins *bs, char *key, char *val);
 void	dup_array_to_env(t_builtins *bs, char **array);
-int		key_checker(t_builtins *bs, char *key, char *val, int pos);
 char	*get_key(t_builtins *bs, int pos);
 char	*get_val(t_builtins *bs, int pos);
-void	mod_key(t_builtins *bs, char *key, char *val);
+void	add_key(t_builtins *bs, char *key, char *val);
 
 //b_unset.c
 int 	b_unset(t_builtins *bs);
@@ -83,5 +74,9 @@ void	remove_key(t_builtins *bs, char *key);
 //b_echo.c
 void    print_echo_n(t_builtins *bs, int i);
 int		b_echo(t_builtins *bs);
+
+//b_init.c
+void	b_init(t_builtins *bs, int argc, char *argv[], char *envp[]);
+void	malloc_checker(char *s);
 
 #endif
