+++ src/builtins.c
@@ -6,45 +6,11 @@
 /*   By: nchennaf <nchennaf@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/24 16:40:41 by nchennaf          #+#    #+#             */
-/*   Updated: 2022/08/11 16:43:37 by nchennaf         ###   ########.fr       */
+/*   Updated: 2022/08/11 16:55:45 by nchennaf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/minishell.h"
-
-// Stocker temporairement les argv entres dans une structure
-// et decaler afin de ne plus avoir le nom "./minishell" en premier
-// argument
-void	b_init(t_builtins *bs, int argc, char *argv[], char *envp[])
-{
-	int	i;
-
-	bs->args = malloc(sizeof(char *) * (argc + 1));
-	malloc_checker((char *)bs->args);
-	// if (!bs->args)
-	// 	exit(EXIT_FAILURE);
-	i = 0;
-	while (envp[i] != NULL)
-		i++;
-	bs->n_env = i;
-	bs->env = malloc(sizeof(char *) * bs->n_env);
-	malloc_checker((char *)bs->env);
-	// if (!bs->env)
-	// 	exit(EXIT_FAILURE);
-	i = 0;
-	while (i < bs->n_env)
-	{
-		bs->env[i] = ft_strdup(envp[i]);
-		i++;
-	}
-	i = 0;
-	while (i < (argc - 1))
-	{
-		bs->args[i] = argv[i + 1];
-		i++;
-	}
-	bs->n_args = i;
-}
 
 // 'pwd'		-> affiche le chemin actuel, suivi d'un \n
 // 'pwd texte	-> message d'erreur : 'pwd: too many arguments'
@@ -110,15 +76,6 @@ int	b_env(t_builtins *bs)
 
 	i = 0;
 	while (i < bs->n_env)
-	{
-		printf("%s\n", bs->env[i]);
-		i++;
-	}
+		printf("%s\n", bs->env[i++]);
 	return (EXIT_SUCCESS);
-}
-
-void	malloc_checker(char *s)
-{
-	if (!s)
-		exit(EXIT_FAILURE);
 }
